Remove selector requirements for getErrorMessages/AsMap
getErrorMessages() and getErrorMessagesAsMap() was not consistent with the rest of the application since it was forcing the user to supply a selector rather than the default behavior of finding all the inputs on the page.

Only getErrorMessage() requires a selector since it used to get an error message for a single particular input.

Resolves #72